refactor: renomeia modulo propertyFromCookie para Cookie
