+++ src/features/common/shared/auth/modules/Cookie/__tests__/propertyFromCookie.test.ts
